# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VigiaApp API
  version: 1.1.0
  description: |
    API REST de VigiaApp. Parte de la especificación v1.0 (`VigiaApp/docs/especificacion_api_db.md`);
    las diferencias con esa versión están en `docs/CAMBIOS_CLIENTE.md`.

    - Todas las respuestas son `application/json; charset=utf-8` (RNF-04).
    - Claves en `snake_case`; los booleanos viajan como `true`/`false` (RNF-07).
    - Fechas con hora: ISO-8601 en UTC con `Z` y sin milisegundos (`2026-09-23T18:35:00Z`).
      Fechas de calendario: `AAAA-MM-DD`. Horas: `HH:MM`.
    - Todas las rutas, salvo `/auth/login`, `/auth/clave` y `/auth/recuperar`, exigen `Authorization: Bearer <token>`.
    - Cualquier respuesta que no sea 2xx trae el cuerpo `Error`.
servers:
  - url: https://{dominio}/api/v1
    variables:
      dominio:
        default: api.vigiaapp.example
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
security:
  - bearer: []
...
paths:
...
  /publicaciones:
    get:
      tags: [Publicaciones]
      summary: Feed de actividad (RF-05) o noticias del comité (RF-06)
      description: Últimas 50, de la más nueva a la más vieja. El listado no trae `cuerpo`.
      parameters:
        - { name: tipo, in: query, required: true, schema: { $ref: '#/components/schemas/TipoPublicacion' } }
        - { name: etiqueta, in: query, required: false, schema: { $ref: '#/components/schemas/Etiqueta' } }
      responses:
        '200':
          description: Publicaciones
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Publicacion' } }
              example:
                - { id: 1, tipo: actividad, titulo: Cierre de la calle Juárez por obra, etiqueta: null, autor: Jorge L., iniciales: JL, autor_id: 3, creado_en: '2026-09-23T18:35:00Z', respuestas: 0 }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
    post:
      tags: [Publicaciones]
      summary: Publicar
      description: Cualquier vecino publica `actividad` (sin etiqueta). `noticia` (con etiqueta obligatoria) y `firma` son solo del comité.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tipo, titulo]
              properties:
                tipo: { $ref: '#/components/schemas/TipoPublicacion' }
                titulo: { type: string, maxLength: 140 }
                cuerpo: { type: [string, 'null'], maxLength: 5000 }
                etiqueta: { oneOf: [{ $ref: '#/components/schemas/Etiqueta' }, { type: 'null' }] }
                firma: { type: [string, 'null'], maxLength: 40, description: 'p. ej. "Comité", "Municipio"' }
      responses:
        '201':
          description: Creada
          content: { application/json: { schema: { $ref: '#/components/schemas/PublicacionDetalle' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
  /publicaciones/{id}:
    get:
      tags: [Publicaciones]
      summary: Una publicación con su cuerpo
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Publicación
          content: { application/json: { schema: { $ref: '#/components/schemas/PublicacionDetalle' } } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    delete:
      tags: [Publicaciones]
      summary: Borrar (el autor o el comité)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Borrada junto con sus respuestas. }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /publicaciones/{id}/respuestas:
    get:
      tags: [Publicaciones]
      summary: Respuestas de una publicación (hasta 200, de la más vieja a la más nueva)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Respuestas
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Respuesta' } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    post:
      tags: [Publicaciones]
      summary: Responder
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [texto]
              properties:
                texto: { type: string, maxLength: 500 }
      responses:
        '201':
          description: Respuesta creada
          content: { application/json: { schema: { $ref: '#/components/schemas/Respuesta' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

  # ---------------------------------------------------------------- Rondines
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: 'Token de 64 hex que devuelven `/auth/login` y `/auth/clave`. Vale 24 h (RNF-03).'
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
...
    DatosInvalidos:
      description: Falta un campo o tiene formato incorrecto.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: datos_invalidos, mensaje: 'presente debe ser true o false.' } }
    NoAutorizado:
      description: Token ausente, inválido o vencido (o credenciales incorrectas en /auth/login).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_autorizado, mensaje: 'Sesión vencida. Volvé a iniciar sesión.' } }
    Prohibido:
      description: El vecino no tiene permiso (p. ej. no es del comité).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: prohibido, mensaje: 'Solo el comité puede hacer esto.' } }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_encontrado, mensaje: 'La alerta no existe.' } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [codigo, mensaje]
          properties:
            codigo:
              type: string
              enum: [datos_invalidos, no_autorizado, prohibido, no_encontrado, conflicto, demasiados_intentos, error_interno]
            mensaje: { type: string, description: 'Texto en castellano, apto para mostrar al usuario.' }
...
    TipoPublicacion:
      type: string
      enum: [actividad, noticia]
    Etiqueta:
      type: string
      enum: [obras, seguridad, alumbrado, convivencia]
    Publicacion:
      type: object
      description: Sirve a `ActivityItem.fromJson` y a `NewsItem.fromJson` (campos extra se ignoran).
      required: [id, tipo, titulo, etiqueta, autor, iniciales, autor_id, creado_en, respuestas]
      properties:
        id: { type: integer }
        tipo: { $ref: '#/components/schemas/TipoPublicacion' }
        titulo: { type: string }
        etiqueta: { oneOf: [{ $ref: '#/components/schemas/Etiqueta' }, { type: 'null' }] }
        autor: { type: string, example: Jorge L., description: 'La `firma` ("Comité") o "Nombre I."' }
        iniciales: { type: string, example: JL }
        autor_id: { type: integer, description: Para saber si el vecino puede borrarla }
        creado_en: { type: string, format: date-time }
        respuestas: { type: integer }
    PublicacionDetalle:
      allOf:
        - $ref: '#/components/schemas/Publicacion'
        - type: object
          required: [cuerpo]
          properties:
            cuerpo: { type: [string, 'null'] }
    Respuesta:
      type: object
      required: [id, texto, autor, iniciales, autor_id, creado_en]
      properties:
        id: { type: integer }
        texto: { type: string }
        autor: { type: string, example: María R. }
        iniciales: { type: string, example: MR }
        autor_id: { type: integer }
        creado_en: { type: string, format: date-time }
...